% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_articles}
\alias{cc_articles}
\title{Hybrid OA publishing output from first authors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 884736 rows and 7 columns.
}
\usage{
cc_articles
}
\description{
This dataset comprises affiliation information from first authors for each
open access article published in hybrid journals as listed in
\link[hoaddata]{oam_hybrid_jns}.
}
\details{
A first author is often considered as lead author (\url{https://en.wikipedia.org/wiki/Lead_author})
who has usually undertaken most of the research presented in the article,
although author roles can vary across disciplines.
We used OpenAlex as data source for determining the lead author's affiliation.

In case where OpenAlex did not record an country affiliation, we extracted
country names from the display_name using regular expressions.

Variables:

\describe{
    \item{doi}{DOI for the OA article}
    \item{issn_l}{Linking ISSN}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{cc}{Normalized Creative Commons variant.}
    \item{country_code}{The country where this institution is located, represented as an ISO two-letter country code. (OpenAlex field `country_code` and extra country extraction)}
    \item{id}{The OpenAlex ID for this institution}
    \item{display_name}{The primary name of the institution (OpenAlex field `display_name`)}
    }

 To our knowledge, OpenAlex does not provide information about corresponding
 authors and their affiliation.
}
\examples{
# Hybrid OA articles with lead author from Uni Göttingen
cc_articles[cc_articles$id \%in\% "https://openalex.org/I74656192",]
}
\keyword{articledatasets}
