% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_jn_ind}
\alias{cc_jn_ind}
\title{Prevalence of Creative Commons licenses by variant, year and journal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 39483 rows and 6 columns.
}
\usage{
cc_jn_ind
}
\description{
This dataset contains the number and proportion of articles with Creative
Commons license (CC) by license variant and year for hybrid journals as
listed by the German Open Access monitor since 2013.
}
\details{
Journal's article volume was calculated using the Crossref database snapshot.
Note that only articles published in regular issues aside from supplements
containing conference contributions like meeting abstracts,
indicated by non-numeric pagination were included. Also, non-scholarly
journal content, such as the table of contents were excluded,
following Unpaywall's paratext recognition approach, which was expanded to
include patterns indicating corrections.

CC licenses were also identified through Crossref metadata records.
License information for author accepted manuscripts ("aam") and
with time lag between licensing and publication (delayed OA) were not
considered.

#' @format A data frame with 3 variables
  \describe{
    \item{issn_l}{Linking International Standard Serial Number (ISSN-L), an umbrella ID for all media versions of the journal}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{cc}{Normalized Creative Commons variant}
    \item{cc_total}{Number of articles under Creative Commons variant}
    \item{jn_all}{Yearly journal output}
    \item{prop}{Proportion of CC licensed articles}
    }
}
\keyword{datasets}
