% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_md_indicators}
\alias{cc_md_indicators}
\title{Crossref Metadata Coverage}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 104 rows and 9 columns.
}
\usage{
cc_md_indicators
}
\description{
This dataset gives information about metadata coverage per year and agreement.
}
\details{
The following Crossref metadata were analysed:

#' \describe{
    \item{tdm_total}{The number of articles containing full text URLs in the metadata}
    \item{orcid_total}{The number of articles containing at least one ORCID in the metadata}
    \item{funder_total}{The number of articles with funding metadata}
    \item{abstract_total}{The number of articles with open abstracts}
    \item{ref_total}{The number of articles with open reference lists)}
    }
}
\examples{
# Wiley metadata coverage
cc_md_indicators[cc_md_indicators$agreement == "Wiley Hybrid",]
}
\keyword{articledatasets}
