% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_openalex_inst_jn_ind}
\alias{cc_openalex_inst_jn_ind}
\title{First author country affiliations per journal, year and Creative Commons license}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 481506 rows and 7 columns.
}
\usage{
cc_openalex_inst_jn_ind
}
\description{
This dataset contains the number and proportion of open access articles in
hybrid journals by country, year and Creative Commons license variant.
}
\details{

}
\examples{
cc_openalex_inst_jn_ind
}
\keyword{jndatasets}
