% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_md}
\alias{cr_md}
\title{Crossref Metadata Coverage}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64497 rows and 9 columns.
}
\usage{
cr_md
}
\description{
This dataset gives information about metadata coverage per year and agreement.
}
\details{
The following Crossref metadata were analysed:

\describe{
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{issn_l}{Linking ISSN}
    \item{articles_total}{Yearly journal output by year and country affiliation}
    \item{tdm_total}{The number of articles containing full text URLs in the metadata}
    \item{orcid_total}{The number of articles containing at least one ORCID in the metadata}
    \item{funder_total}{The number of articles with funding metadata}
    \item{abstract_total}{The number of articles with open abstracts}
    \item{ref_total}{The number of articles with open reference lists)}
    \item{cat}{Global output or from first-authors based in Germany}
    }
}
\examples{
# OA Articles in Scientometrics from first-authors in Germany: Metadata coverage
cr_md[cr_md$issn_l == "0138-9130" & cr_md$cat == "Germany",]
}
\keyword{articledatasets}
