% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_cc_inst}
\alias{jct_cc_inst}
\title{JCT Hybrid Journals: Article-level data first author affiliation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 799182 rows and 7 columns.
}
\usage{
jct_cc_inst
}
\description{
JCT Hybrid Journals: Article-level data first author affiliation
}
\keyword{datasets}
