% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_cr_md}
\alias{jct_cr_md}
\title{JCT Hybrid Journals: Crossref Metadata Coverage}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 411 rows and 9 columns.
}
\usage{
jct_cr_md
}
\description{
JCT Hybrid Journals: Crossref Metadata Coverage
}
\keyword{datasets}
