% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_cr_upw}
\alias{jct_cr_upw}
\title{JCT Hybrid Journals: License Coverage}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 470 rows and 6 columns.
}
\usage{
jct_cr_upw
}
\description{
JCT Hybrid Journals: License Coverage
}
\keyword{datasets}
