% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_jn_aff}
\alias{jct_jn_aff}
\title{JCT Hybrid Journals: First Author Country Affiliation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1497816 rows and 6 columns.
}
\usage{
jct_jn_aff
}
\description{
JCT Hybrid Journals: First Author Country Affiliation
}
\keyword{datasets}
