% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_jn_ind}
\alias{jct_jn_ind}
\title{JCT Hybrid Journals: Article Volume}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 94496 rows and 6 columns.
}
\usage{
jct_jn_ind
}
\description{
JCT Hybrid Journals: Article Volume
}
\keyword{datasets}
