% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_oalex_venues}
\alias{jct_oalex_venues}
\title{Hybrid Journals: Venues Metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11839 rows and 4 columns.
}
\usage{
jct_oalex_venues
}
\description{
Hybrid Journals: Venues Metadata
}
\keyword{datasets}
