% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jn_ind}
\alias{jn_ind}
\title{Prevalence of Creative Commons licenses by variant, year and journal}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 99575 rows and 6 columns.
}
\usage{
jn_ind
}
\description{
This dataset contains the number and proportion of open access articles
with Creative Commons license (CC) by license variant and year for hybrid journals
included in the cOAlition S Journal Checker Tool.
}
\details{
Publication period is 2017 - 2022.

Journal's article volume was calculated using Crossref metadata snapshot.
Note that only articles published in regular issues aside from supplements
containing conference contributions like meeting abstracts,
indicated by non-numeric pagination, were included. Also, non-scholarly
journal content, such as the table of contents were excluded. In doing so,
we followed Unpaywall's paratext recognition approach, which we expanded to
include patterns indicating corrections.

CC licenses were also identified through Crossref.
License information for author accepted manuscripts ("aam") were not
considered.

Variables:

\describe{
    \item{issn_l}{Linking ISSN}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{cc}{Normalized Creative Commons variant. Ordered factor by license variant permissiveness}
    \item{cc_total}{Number of articles under Creative Commons variant}
    \item{jn_all}{Yearly journal output}
    \item{prop}{Proportion of CC licensed articles}
    }
}
\examples{
jn_ind
}
\keyword{jndatasets}
