% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oam_df}
\alias{oam_df}
\title{Hybrid Journals provided by the Open Access Monitor}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10769 rows and 3 columns.
}
\usage{
oam_df
}
\description{
A dataset comprising hybrid journals from German transformative agreements.
The dataset was unified and enriched with ISSN variants and ISSN-L.
}
\keyword{datasets}
