% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oam_hybrid_jns}
\alias{oam_hybrid_jns}
\title{Hybrid Journals listed in the Open Access Monitor}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11638 rows and 4 columns.
}
\source{
Pollack, Philipp; Lindstrot, Barbara; Barbers, Irene,
Stanzel, Franziska 2022, Open Access Monitor: Zeitschriftenlisten (v2)
\url{https://doi.org/10.26165/JUELICH-DATA/VTQXLM}
}
\usage{
oam_hybrid_jns
}
\description{
The dataset contains hybrid journals available as a filter in the
\href{https://open-access-monitor.de/}{German Open Access Monitor}.
The dataset was unified and mapped to the
\href{https://www.crossref.org/titleList/}{Crossref title list}.
}
\details{
There's no direct data exchange between Crossref and the ISSN agency.
Instead, publishers register journal-level metadata when they first deposit
metadata for a given journal, including all ISSN(s). Crossref makes sure that
article and journal metadata match.

More info about Crossref's handling of ISSN registration can be found in this
support thread:
\url{https://community.crossref.org/t/parallel-titles-for-a-given-issn/2183}


Variables:

\describe{
    \item{agreement}{Transformative agreement from a German consortium}
    \item{lead}{Institution leading the nationwide consortium}
    \item{cr_journal_id}{Crossref journal ID}
    \item{issn}{International Standard Serial Number (ISSN), a ID to refer to a specific journal's media version}
    }
}
\examples{
oam_hybrid_jns
}
\keyword{jndatasets}
