% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oam_jn_by_year}
\alias{oam_jn_by_year}
\title{Yearly publication volume}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46659 rows and 3 columns.
}
\usage{
oam_jn_by_year
}
\description{
Yearly publication volume
}
\keyword{datasets}
