% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subugue_bq_auth}
\alias{subugue_bq_auth}
\title{Helper function ----
Authentificate with Big Query}
\usage{
subugue_bq_auth()
}
\description{
Decrypting-ability using internal secret_* functions from gargle
}
