% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ta_country_output}
\alias{ta_country_output}
\title{JCT Hybrid Journals: Institutions involved in transformative agreements by country}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1396711 rows and 6 columns.
}
\usage{
ta_country_output
}
\description{
Variables: 
\describe{
    \item{issn_l}{Linking ISSN}				
    \item{country_code}{The country where this institution is located, represented as an ISO two-letter country code.}				
    \item{cr_year}{Earliest publication year (Crossref field `issued`)} 			
    \item{oa_n}{Number of distinct open access articles}				
    \item{n}{Number of distinct articles}			
    \item{has_ta}{Involvement in transformative agreements}
    }
}
\examples{
# Article numbers (total `n` and open access `oa_n`) for journals involved in transformative agreements in Germany
subset(ta_country_output, country_code == "DE" & has_ta == TRUE)
}
\keyword{datasets}
