% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_openalex_inst_jn_ind}
\alias{cc_openalex_inst_jn_ind}
\title{First author country affiliations per journal, year and Creative Commons license}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 633122 rows and 6 columns.
}
\usage{
cc_openalex_inst_jn_ind
}
\description{
This dataset contains the number and proportion of open access articles in
hybrid journals by country, year and Creative Commons license variant.
}
\details{
Country affiliations were determined using the first author affiliation as
listed by OpenAlex.
A first author is often considered as lead
author(\url{https://en.wikipedia.org/wiki/Lead_author})
who has usually undertaken most of the research presented in the article,
although author roles can vary across disciplines.

Note that whole counting in which every original article or review was
counted once per country of affiliation of the first authors were used.
Because first authors can have multiple affiliations from different countries,
please use \code{\link{cc_jn_ind}} to determine a journal's publication
volume.

Variables:
\describe{
    \item{issn_l}{Linking International Standard Serial Number (ISSN-L), an umbrella ID for all media versions of the journal}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{country_code}{The country where this institution is located, represented as an ISO two-letter country code. (OpenAlex field `country_code`)}
    \item{cc}{Normalized Creative Commons variant. \code{NA} represents articles, which were not provided under a CC license}
    \item{articles_under_cc_variant}{Number of articles under Creative Commons variant, grouped by journal, year and country affiliation}
    \item{articles_total}{Yearly journal output by year and country affiliation}
    }
}
\examples{
  subset(cc_openalex_inst_jn_ind,
    issn_l \%in\% "0138-9130" & country_code \%in\% c("DE", "CN"))
}
\keyword{jndatasets}
