% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_openalex_inst_jn_ind}
\alias{cc_openalex_inst_jn_ind}
\title{First author country affiliations per journal, year and Creative Commons license}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 481506 rows and 7 columns.
}
\usage{
cc_openalex_inst_jn_ind
}
\description{
This dataset contains the number and proportion of open access articles in
hybrid journals by country, year and Creative Commons license variant.
}
\details{
Country affiliations were determined using the first author affiliation as
listed by OpenAlex.
A first author is often considered as lead
author(\url{https://en.wikipedia.org/wiki/Lead_author})
who has usually undertaken most of the research presented in the article,
although author roles can vary across disciplines.

Variables:
\describe{
    \item{issn_l}{Linking International Standard Serial Number (ISSN-L), an umbrella ID for all media versions of the journal}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{country_code}{The country where this institution is located, represented as an ISO two-letter country code. (OpenAlex field `country_code`)}
    \item{cc}{Normalized Creative Commons variant.}
    \item{cc_total}{Number of articles under Creative Commons variant by journal, year and country affiliation}
    \item{jn_all}{Yearly journal output by year and country affiliation}
    \item{prop}{Proportion of articles under CC license variant by journal, year and country affiliation}
    }
}
\examples{
cc_openalex_inst_jn_ind
}
\keyword{jndatasets}
