% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_upw}
\alias{cr_upw}
\title{License coverage Crossref vs Unpaywall}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 91889 rows and 6 columns.
}
\usage{
cr_upw
}
\description{
This dataset provides the data for the comparision of license metadata in Crossref with Unpaywall per year.
}
\details{
Variables:
\describe{
    \item{issn_l}{Linking ISSN}
    \item{cr_year}{Earliest publication year (Crossref field `issued`)}
    \item{articles_total}{Yearly journal output by year}
    \item{upw_hybrid_total}{Total number of unpaywall hybrid articles}
    \item{cr_hybrid_total}{Total number of crossref hybrid articles}
    \item{cat}{Global output or from first-authors based in Germany}
    }
}
\examples{
# Number of crossref hybrid articles vs all unpaywall hybrid articles (Global)
sum(subset(cr_upw, cat == "Germany"& cr_year == 2020)$upw_hybrid_total)
sum(subset(cr_upw, cat == "Germany"& cr_year == 2020)$cr_hybrid_total)
# Number of crossref hybrid articles vs all unpaywall hybrid articles (Germany)
sum(subset(cr_upw, cat == "Global" & cr_year == 2020)$upw_hybrid_total)
sum(subset(cr_upw, cat == "Global" & cr_year == 2020)$cr_hybrid_total)
}
\keyword{datasets}
