% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_upw}
\alias{cr_upw}
\title{License coverage Crossref vs Unpaywall}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92018 rows and 6 columns.
}
\usage{
cr_upw
}
\description{
License coverage Crossref vs Unpaywall
}
\examples{
cr_upw
}
\keyword{datasets}
