% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_upw}
\alias{cr_upw}
\title{License coverage Crossref vs Unpaywall}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110223 rows and 6 columns.
}
\usage{
cr_upw
}
\description{
Comparision of license metadata in Crossref with Unpaywall
via OpenALEX per journal, year and country affiliation (Germany).
}
\examples{
cr_upw
}
\keyword{datasets}
