% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_hybrid_jns}
\alias{jct_hybrid_jns}
\title{Hybrid Journals}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 384905 rows and 4 columns.
}
\usage{
jct_hybrid_jns
}
\description{
Hybrid Journals
}
\examples{
jct_hybrid_jns
}
\keyword{datasets}
