% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jct_oalex_venues}
\alias{jct_oalex_venues}
\title{Hybrid Journals: Venues Metadata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11820 rows and 4 columns.
}
\usage{
jct_oalex_venues
}
\description{
This dataset gives information on the venue of hybrid journals providing an url link to journals that are listed with their issn number, OpenAlex id and the name.
}
\details{
Variables:
\describe{
    \item{issn_l}{Linking ISSN}
    \item{id}{The OpenAlex ID for this institution}
    \item{display_name}{Journal name}
    \item{homepage_url}{Journal url}
    }
}
\examples{
# Venues metadata for an issn_l
jct_oalex_venues[jct_oalex_venues$issn_l == "0006-3568",]
}
\keyword{datasets}
