% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ta_country_output}
\alias{ta_country_output}
\title{JCT Hybrid Journals: Institutions involved in transformative agreements by country}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1615399 rows and 6 columns.
}
\usage{
ta_country_output
}
\description{
JCT Hybrid Journals: Institutions involved in transformative agreements by country
}
\examples{
ta_country_output
}
\keyword{datasets}
